[mod]一些win下链接库的代码加了一个条件编译的控制

+++ src/WtExecMon/WtExecPorter.h
@@ -22,7 +22,7 @@ extern "C"
 
 	EXPORT_FLAG	void		run_exec();
 
-	EXPORT_FLAG	void		write_log(unsigned int level, WtString message, WtString catName);
+	EXPORT_FLAG	void		write_log(unsigned long level, WtString message, WtString catName);
 
 	EXPORT_FLAG	WtString	get_version();
